scripts/requirements.txt
garminconnect>=0.2.40
garth>=0.4
supabase>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
browser-cookie3>=0.19.0
psycopg2-binary>=2.9.0
beautifulsoup4>=4.12.0
